+++ CHANGELOG.md
@@ -3,12 +3,15 @@
 ### Features
 
 * Native integration with chatwoot
+* Added returning or non-returning participants option in fetchAllGroups
+
 ### Fixed
 
 * Adjusts in docker-compose files
 * Adjusts in number validation for AR and MX numbers
 * Adjusts in env files, removed save old_messages
 * Fix when sending a message to a group I don't belong returns a bad request
+* Fits the format on return from the fetchAllGroups endpoint
 
 # 1.1.5 (2023-07-12 07:17)